fixed one problem, created another
