+++ src/components/design/DesignList.js
@@ -12,6 +12,7 @@ const DesignList = props => {
     const getFollowedUserDesigns = () => {
         if (followedUser) {
             //get the other users designs
+            console.log(props.followedUserId)
             getOtherUserDesigns(props.match.params.userId)
                 .then(setFollowedDesigns)
         }
