+++ src/components/providers/DesignProvider.js
@@ -13,6 +13,7 @@ export const DesignContext = React.createContext()
 export const DesignProvider = props => {
     const [designs, setDesigns] = useState([])
     const [currentUser] = localStorage.getItem("currUserId")
+    console.log("DesignProvider", currentUser)
 
 
     const getDesigns = () => {
